Refactored in package with exercise number
