+++ leertaak1/src/main/java/edited/opdracht8/DobbelsteenModel.java
@@ -1,4 +1,4 @@
-package edited;
+package edited.opdracht8;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
