+++ leertaak1/src/main/java/unedited/opdracht8/DobbelsteenModel.java
@@ -1,4 +1,4 @@
-package unedited;
+package unedited.opdracht8;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
